ExprAttributeModifierOfItem - update to include entities
